requirements.txt
pandas==1.2.4
numpy==1.20.2
PyYAML==5.4.1
scikit_learn==0.24.2
